+++ resources/js/views/FormNested.vue
@@ -398,7 +398,7 @@
             needsReload() {
                 return JSON.stringify({
                     defaultChildren: this.defaultChildren,
-                    propagated: this.propagated,
+                    propagated: this.watchablePropagated,
                     hiddenFields: this.hiddenFields,
                 });
             },